change to raw and seq module names
